Converted project to gradle and update to latest minecraft
Removed eclipse project files and added gradle layout
Repackaged for simplemc
Updated to Spigot for MC 1.13
Fixed deprecated API calls
Fixed minor code smells
