--- .classpath
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="var" path="BUKKIT" sourcepath="/BUKKIT_SOURCE"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
