+++ build.gradle
@@ -1,0 +1,25 @@
+plugins {
+    id 'java'
+    id 'maven'
+}
+
+group 'org.simplemc'
+version '1.13-SNAPSHOT'
+
+// target Java 8
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+
+// enforce UTF-8
+compileJava.options.encoding = 'UTF-8'
+javadoc.options.encoding = 'UTF-8'
+
+repositories {
+    jcenter()
+    maven { url "https://hub.spigotmc.org/nexus/content/repositories/snapshots/" } // spigot
+    maven { url "https://oss.sonatype.org/content/repositories/snapshots" } // bungeecord
+}
+
+dependencies {
+    compile 'org.spigotmc:spigot-api:1.13.2-R0.1-SNAPSHOT'
+}
