--- plugin.yml
@@ -1,5 +1,0 @@
-name: SimpleHealthbars
-main: com.evosysdev.bukkit.taylorjb.simplehealthbars.SimpleHealthbars
-version: 0.0.1
-website: https://github.com/Evosysdev/SimpleHealthbars
-author: taylorjb
